Fixes Stratagem Ball block placement
Fixes an issue where Stratagem Ball was using incorrect block position to place block.
Also fixes Stratagem Pod logic to prevent block removal issue.

+++ src/main/java/com/stevekung/stratagems/entity/StratagemPod.java
@@ -70,7 +70,7 @@ public class StratagemPod extends Entity implements VariantHolder<Holder<Stratag
 
         if (!this.level().isClientSide() && this.inboundTick == 0 && this.getOwner() instanceof ServerPlayer serverPlayer)
         {
-            if (this.linkedBlock != null)
+            if (this.linkedBlock != null && this.linkedBlockPos != null)
             {
                 this.level().destroyBlock(this.linkedBlockPos, false);
                 this.level().sendBlockUpdated(this.linkedBlockPos, this.linkedBlock.defaultBlockState(), Blocks.AIR.defaultBlockState(), 3);
